added weather, footsteps, tiled map start of textbubbles and day night cycle

+++ src/org/elmarsoft/main/Main.java
@@ -3,13 +3,13 @@ package org.elmarsoft.main;
 import org.newdawn.slick.AppGameContainer;
 import org.newdawn.slick.SlickException;
 
-public class Main{
+public class Main {
 
     public static void main(String[] args) {
         try {
             Game game = new Game("RPG");
             AppGameContainer app = new AppGameContainer(game);
-            app.setDisplayMode(Game.SCREENWIDTH, Game.SCREENHEIGTH, false);
+            app.setDisplayMode((int) Game.SCREENWIDTH, (int) Game.SCREENHEIGTH, false);
             app.setAlwaysRender(true);
             app.setShowFPS(true);
             app.start();
